fix: Fix decimal is getting dropped

+++ eplk/src/main/java/com/iyxan23/eplk/objects/EplkInteger.kt
@@ -92,7 +92,7 @@ class EplkInteger(
 
         return when (other) {
             is EplkInteger ->
-                RealtimeResult<EplkObject>().success(EplkFloat((value / other.value).toFloat(), scope))
+                RealtimeResult<EplkObject>().success(EplkFloat(value.toFloat() / other.value.toFloat(), scope))
 
             is EplkFloat ->
                 RealtimeResult<EplkObject>().success(EplkFloat(value / other.value, scope))
